fix findsnpatpos bug

+++ src/BamProcess.h
@@ -78,7 +78,7 @@ class BamProcess: public SeqLib::BamReader
 
     int GetOffset(const SeqLib::BamRecord& r, const uint32_t pos) const;
 
-    std::unordered_map<char, int> base_m{ {'A', 0},{'C', 1},{'G', 2},{'T', 3},{'N', 4} };
+    std::unordered_map<char, int8_t> base_m{ {'A', 0},{'C', 1},{'G', 2},{'T', 3},{'N', 4} };
 };
